test addMember

+++ nodejs/app.js
@@ -219,13 +219,17 @@ app.post('/addMember', function (req, res) {
         const sourcePath = path.join(__dirname, '..', 'members', 'user-images', 'temp', username);
         const destPath = path.join(__dirname, '..', 'members', 'user-images', username);
 
-        fs.rename(sourcePath, destPath, (err) => {
-            if (err) {
-                console.error(`Error moving file: ${err}`);
-            } else {
-                console.log(`File moved successfully from ${sourcePath} to ${destPath}`);
-            }
-        });
+        if (fs.existsSync(sourcePath)) {
+            fs.rename(sourcePath, destPath, (err) => {
+                if (err) {
+                    console.log(`Error moving file: ${err}`);
+                } else {
+                    console.log(`File moved successfully from ${sourcePath} to ${destPath}`);
+                }
+            });
+        } else {
+            console.log(`Error: Source file does not exist: ${sourcePath}`);
+        }
 
     } else {
         res.send({ "error": true, "message": 'Invalid Email' });
